Create general function to calculate top group and active years
