+++ client/src/panels/panel_declarations/people/employee_age.yaml
@@ -6,23 +6,23 @@ employee_age_title:
 gov_employee_age_text:
   transform: [handlebars,markdown]
   en: |
-   From **{{ppl_last_year_5}}** to **{{ppl_last_year}}**, people aged **{{fmt_big_int top_age_group}}** accounted for the largest average share (**{{fmt_percentage1 top_age_group_pct}}**) of **Federal Public Service** employees.
+   From **{{ppl_last_year_5}}** to **{{ppl_last_year}}**, people aged **{{fmt_big_int top_group}}** accounted for the largest average share (**{{fmt_percentage1 top_group_pct}}**) of **Federal Public Service** employees.
    
    Over this period, the **average age** of **Federal Public Service** employees **{{two_value_changed_to gov_avgage_last_year_5 gov_avgage_last_year "fmt_decimal1"}}**.
   fr: |
-   De **{{ppl_last_year_5}}** à **{{ppl_last_year}}**, les personnes âgées de **{{fmt_big_int top_age_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_age_group_pct}}**) du personnel de la **fonction publique fédérale**.
+   De **{{ppl_last_year_5}}** à **{{ppl_last_year}}**, les personnes âgées de **{{fmt_big_int top_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_group_pct}}**) du personnel de la **fonction publique fédérale**.
 
    Au cours de cette période, **l’âge moyen** des employés de la **fonction publique fédérale** **{{fr_two_value_changed_to gov_avgage_last_year_5 gov_avgage_last_year "m" "s" "fmt_decimal1"}}**.
 
 dept_employee_age_text:
   transform: [handlebars,markdown]
   en: |
-   From **{{fmt_big_int first_active_year}}** to **{{fmt_big_int last_active_year}}**, people aged **{{fmt_big_int top_age_group}}** accounted for the largest average share (**{{fmt_percentage1 top_age_group_pct}}**) of **{{subject.name}}** employees.
+   From **{{fmt_big_int first_active_year}}** to **{{fmt_big_int last_active_year}}**, people aged **{{fmt_big_int top_group}}** accounted for the largest average share (**{{fmt_percentage1 top_group_pct}}**) of **{{subject.name}}** employees.
 
    Over this period, the **average age** of **{{subject.name}}** employees **{{two_value_changed_to dept_avg_first_active_year dept_avg_last_active_year "fmt_decimal1"}}**.
    For comparison, the average age of all **Federal Public Service** employees **{{two_value_changed_to gov_avgage_last_year_5 gov_avgage_last_year "fmt_decimal1"}}**.
   fr: |
-   De **{{fmt_big_int first_active_year}}** à **{{fmt_big_int last_active_year}}**, les personnes âgées de **{{fmt_big_int top_age_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_age_group_pct}}**) d’employés **{{de_dept dept}}**.
+   De **{{fmt_big_int first_active_year}}** à **{{fmt_big_int last_active_year}}**, les personnes âgées de **{{fmt_big_int top_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_group_pct}}**) d’employés **{{de_dept dept}}**.
 
    Au cours de cette période, **l’âge moyen** des employés **{{de_dept dept}}** **{{fr_two_value_changed_to dept_avg_first_active_year dept_avg_last_active_year "m" "s" "fmt_decimal1"}}**.
    Aux fins de comparaison, l’âge moyen de tous les employés de la **fonction publique fédérale** **{{fr_two_value_changed_to gov_avgage_last_year_5 gov_avgage_last_year "m" "s" "fmt_decimal1"}}**.
@@ -30,15 +30,15 @@ dept_employee_age_text:
 gov_employee_age_text_temporary_no_avg_age:
   transform: [handlebars,markdown]
   en: |
-   From **{{ppl_last_year_5}}** to **{{ppl_last_year}}**, people aged **{{fmt_big_int top_age_group}}** accounted for the largest average share (**{{fmt_percentage1 top_age_group_pct}}**) of **Federal Public Service** employees.
+   From **{{ppl_last_year_5}}** to **{{ppl_last_year}}**, people aged **{{fmt_big_int top_group}}** accounted for the largest average share (**{{fmt_percentage1 top_group_pct}}**) of **Federal Public Service** employees.
    
   fr: |
-   De **{{ppl_last_year_5}}** à **{{ppl_last_year}}**, les personnes âgées de **{{fmt_big_int top_age_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_age_group_pct}}**) du personnel de la **fonction publique fédérale**.
+   De **{{ppl_last_year_5}}** à **{{ppl_last_year}}**, les personnes âgées de **{{fmt_big_int top_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_group_pct}}**) du personnel de la **fonction publique fédérale**.
 
 dept_employee_age_text_temporary_no_avg_age:
   transform: [handlebars,markdown]
   en: |
-   From **{{fmt_big_int first_active_year}}** to **{{fmt_big_int last_active_year}}**, people aged **{{fmt_big_int top_age_group}}** accounted for the largest average share (**{{fmt_percentage1 top_age_group_pct}}**) of **{{subject.name}}** employees.
+   From **{{fmt_big_int first_active_year}}** to **{{fmt_big_int last_active_year}}**, people aged **{{fmt_big_int top_group}}** accounted for the largest average share (**{{fmt_percentage1 top_group_pct}}**) of **{{subject.name}}** employees.
 
   fr: |
-   De **{{fmt_big_int first_active_year}}** à **{{fmt_big_int last_active_year}}**, les personnes âgées de **{{fmt_big_int top_age_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_age_group_pct}}**) d’employés **{{de_dept dept}}**.
+   De **{{fmt_big_int first_active_year}}** à **{{fmt_big_int last_active_year}}**, les personnes âgées de **{{fmt_big_int top_group}}** constituaient en moyenne le plus fort contingent (**{{fmt_percentage1 top_group_pct}}**) d’employés **{{de_dept dept}}**.
